+++ src/common/ManagedCommon/RunnerHelper.cs
@@ -3,12 +3,8 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
-using System.Diagnostics;
 using System.Reflection;
 using System.Threading.Tasks;
-
-using Microsoft.PowerToys.Telemetry;
-using Microsoft.PowerToys.Telemetry.Events;
 
 namespace ManagedCommon
 {
@@ -16,9 +12,8 @@ namespace ManagedCommon
     {
         public static void WaitForPowerToysRunner(int powerToysPID, Action act, [System.Runtime.CompilerServices.CallerMemberName] string memberName = "")
         {
-            var stackTrace = new StackTrace();
             var assembly = Assembly.GetCallingAssembly().GetName();
-            PowerToysTelemetry.Log.WriteEvent(new DebugEvent() { Message = $"[{assembly}][{memberName}]WaitForPowerToysRunner waiting for Event powerToysPID={powerToysPID}" });
+            Logger.LogDebug($"[{assembly}][{memberName}]WaitForPowerToysRunner waiting for Event powerToysPID={powerToysPID}");
             Task.Run(() =>
             {
                 const uint INFINITE = 0xFFFFFFFF;
@@ -28,7 +23,7 @@ namespace ManagedCommon
                 IntPtr powerToysProcHandle = NativeMethods.OpenProcess(SYNCHRONIZE, false, powerToysPID);
                 if (NativeMethods.WaitForSingleObject(powerToysProcHandle, INFINITE) == WAIT_OBJECT_0)
                 {
-                    PowerToysTelemetry.Log.WriteEvent(new DebugEvent() { Message = $"[{assembly}][{memberName}]WaitForPowerToysRunner Event Notified powerToysPID={powerToysPID}" });
+                    Logger.LogDebug($"[{assembly}][{memberName}]WaitForPowerToysRunner Event Notified powerToysPID={powerToysPID}");
                     act.Invoke();
                 }
             });