+++ src/settings-ui/Settings.UI/SettingsXAML/MainWindow.xaml.cs
@@ -5,12 +5,10 @@
 using System;
 using System.Threading.Tasks;
 using ManagedCommon;
-using Microsoft.PowerLauncher.Telemetry;
 using Microsoft.PowerToys.Settings.UI.Helpers;
 using Microsoft.PowerToys.Settings.UI.Library;
 using Microsoft.PowerToys.Settings.UI.Library.Helpers;
 using Microsoft.PowerToys.Settings.UI.Views;
-using Microsoft.PowerToys.Telemetry;
 using Microsoft.UI;
 using Microsoft.UI.Dispatching;
 using Microsoft.UI.Windowing;
@@ -25,9 +23,6 @@ namespace Microsoft.PowerToys.Settings.UI
     {
         public MainWindow(bool createHidden = false)
         {
-            var bootTime = new System.Diagnostics.Stopwatch();
-            bootTime.Start();
-
             this.Activated += Window_Activated_SetIcon;
 
             App.ThemeService.ThemeChanged += OnThemeChanged;
@@ -129,10 +124,6 @@ namespace Microsoft.PowerToys.Settings.UI
                     }
                 }
             };
-
-            bootTime.Stop();
-
-            PowerToysTelemetry.Log.WriteEvent(new SettingsBootEvent() { BootTimeMs = bootTime.ElapsedMilliseconds });
         }
 
         private void SetTitleBar()